requirements.txt
beautifulsoup4==4.12.2
pymongo==4.0.2
requests==2.27.1
python-dotenv==1.0.0
soupsieve==2.4.1
urllib3==2.0.3
dnspython==2.3.0
certifi==2023.5.7
idna==3.4
charset-normalizer==3.1.0